requirements.txt
pyln-client>=0.12
pyln-testing
